API, timeline and auto update to the latest version

+++ src/assets/js/modules/config.js
@@ -3,6 +3,40 @@
  */
 $(function(){
 
+	// Gestion de la version -----------------------------------
+	$.get('version', (version) => {
+		const VERSION = version;
+
+		if (localStorage.getItem("appUpdated") == VERSION) {
+			showToastUpdate("done");
+			localStorage.removeItem("appUpdated");
+		}
+
+		if (localStorage.getItem("version")) {
+			if (VERSION != localStorage.getItem("version")) {
+
+				bootbox.dialog({ 
+				    message: '<div class="text-center"><img src="src/assets/img/loader-info.svg" width="24px"> Mise à jour...</div>', 
+				    closeButton: false 
+				});
+
+				caches.keys().then( keyList => keyList.map(key => caches.delete(key)));
+				localStorage.setItem("version", VERSION);
+				localStorage.setItem("appUpdated", VERSION);
+
+				setTimeout(function(){
+					document.location = '';
+				}, 500);
+			}
+		} else {
+			localStorage.setItem("version", VERSION);
+		}
+
+		$('#version').html(localStorage.getItem("version"));
+	});
+	// ---------------------------------------------------------
+
+
 	// --- Config - Preferences
 	$('.settings .btn-toggle[default-value]')
 	// Enregistre dans le stockage local les paramètres par défaut 
@@ -30,23 +64,57 @@ $(function(){
 
 	// --- Config - Data cached
 	$('#emptyCookies').click(function(){
-		$(this).html("En cours...")
-		localStorage.clear();
-		document.location = '';
+		$('#emptyCookies .loader').show();
+		bootbox.confirm("Les fichiers en cache et toutes les valeurs enregistrées seront effacées.", function(result){
+			if (result) {
+				fetch('src/call_api.php?endpoint=users/'+localStorage.getItem("coupling-code"), {
+	      			method: 'DELETE'
+	  			}).then(response => {
+	  				caches.keys().then( keyList => keyList.map(key => caches.delete(key)));
+	  				localStorage.clear();
+	  				document.location = '';
+	  			});
+			} else {
+				$('#emptyCookies .loader').hide();
+			}
+		});
 	});
 
 	// --- Config - Device coupling 
 	if(!localStorage.getItem("coupling-code")) {
 		localStorage.setItem("coupling-code", Math.random().toString(36).substring(4, 15));
 	}
-	$("#coupling-code").val(localStorage.getItem("coupling-code"));
+	$("input.coupling-code").val(localStorage.getItem("coupling-code"));
+	$("div.coupling-code").html(localStorage.getItem("coupling-code"));
 	
 	$('#renew-coupling-code').click(function(){
 		localStorage.setItem("coupling-code", Math.random().toString(36).substring(4, 15));
-		$("#coupling-code").val(localStorage.getItem("coupling-code"));
+		$("input.coupling-code").val(localStorage.getItem("coupling-code"));
+		$("div.coupling-code").html(localStorage.getItem("coupling-code"));
 	});
 
+	// -- Config - Username
+	if(localStorage.getItem("username")) {
+		$('#config-username').addClass("d-flex").show();
+		$('div.username').html(localStorage.getItem("username"));
+	}
 	
 });
+
+function showToastUpdate(context) {
+    var idLink = (context == "sw") ? "update-sw" : "delete-cache";
+    var content = '<div class="d-flex"><div class="mr-2">Nouvelle version disponible.</div> <a href="#" id="'+idLink+'" class="font-weight-bold">Mettre à jour</a></div>';
+    var toastOptions = {autohide:false};
+
+    if (context == "done") {
+    	content = '<div class="d-flex"><div><i class="material-icons mr-2 text-success" style="vertical-align: bottom;">check_circle_outline</i>Mise à jour installée: <b>Version '+localStorage.getItem("version")+'</b></div></div>';
+    	toastOptions = {delay:10000};
+    }
+
+    $('.toast .toast-header').hide();
+    $('.toast-body').html(content);
+    $('.toast').toast(toastOptions);
+    $('.toast').toast("show");
+}
